fix(02-07): drive make typecheck clean across the D-05 in-scope package (M2-03)
Option-2 scope: clean mypy --strict across the in-scope backtest path + shared
core (events, enums, exceptions, config, portfolio, order, execution, strategy
base/sizers/risk, csv price feed, reporting base); out-of-scope debt deferred via
documented ignore_errors overrides.

- SignalEvent.strategy_id int->StrategyId retype (02-05 carry-over); OrderEvent
  too. SignalEvent left mutable (M3 #11).
- Removed dead local old_status at order.py (Pylance-flagged).
- Documented ignore_errors overrides: my_strategies.*, legacy_config,
  postgresql_storage, config.exchange.schema, reporting.statistics/engine_logger/
  plots, events_handler.screener_event_handler. Third-party stub-ignores: pandas,
  pytz, scipy, plotly, sklearn, statsmodels, tqdm, yaml.
- Behavior-preserving: Decimal/float coerced only at existing M2a boundaries;
  int<->PortfolioId bridged with documented unions (full retype deferred).
- Rule 1 fixes in dormant code: ExecutionErrorCode.TIMEOUT added, super().__init__
  in Empty_strategy, NotImplemented->NotImplementedError, regex None-guards.

make typecheck exits 0 (157 files); make test green (299); oracle green.

+++ itrader/execution_handler/result_objects.py
@@ -7,6 +7,7 @@ following the iTrader system's established patterns for data structures.
 
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional, Dict, Any
 
 from itrader.core.enums.execution import ExecutionStatus, ExecutionErrorCode, ExchangeConnectionStatus
@@ -31,7 +32,8 @@ class ExecutionResult:
     executed_price: Optional[float] = None
     executed_quantity: Optional[float] = None
     remaining_quantity: Optional[float] = None
-    commission: Optional[float] = None
+    # M2a money boundary: commission is the Decimal returned by the fee model.
+    commission: Optional[Decimal] = None
     execution_time: Optional[datetime] = None
     
     # Error information
@@ -69,7 +71,7 @@ class ExecutionResult:
         """Calculate net value after commission."""
         total = self.total_value
         if total is not None and self.commission is not None:
-            return total - self.commission
+            return total - float(self.commission)
         return total
